+++ src/ui/dock.js
@@ -3,9 +3,9 @@
 import { store } from './store.js?v=e9304c5ee2';
 import { h, fmt, icon, svgIcon, popover, closePopover, clamp } from './util.js?v=13768f12bf';
 import { NODES, EDGES } from '../engine/topology.js?v=44e0aca402';
-import { createProfile, createScope, createSankey, createPerfusion } from './charts.js?v=5f03127400';
-import { createHVPG, createDoppler, createEndoscopy, createVarixWall, createAbdomen } from './instruments.js?v=ee67c9a9b1';
-import { createLandscape } from './landscape.js?v=ca9777077f';
+import { createProfile, createScope, createSankey, createPerfusion } from './charts.js?v=af67e3e2ae';
+import { createHVPG, createDoppler, createEndoscopy, createVarixWall, createAbdomen } from './instruments.js?v=4d4f3fb0eb';
+import { createLandscape } from './landscape.js?v=b6a8cf0a62';
 
 const NI = Object.fromEntries(NODES.map((n, i) => [n.id, i]));
 const EI = Object.fromEntries(EDGES.map((e, i) => [e.id, i]));